rocjitsu: fix UBSan-found SIMD overflows + cmp_class/mov semantics
Found via UBSan build of the SIMD fast paths (run with -fsanitize=undefined,
rocjitsu_tests SIMD/vop suite): 2 correctness failures + integer-overflow UB.

Codegen (source of truth):
- packed.py: pk u16 mul and pk int mad widen operands to uint32 before the
  multiply (uint16*uint16 was int-promoted -> signed overflow).
- vector_alu.py: i24 mad computes a*b in int64 before truncation.

Regenerated bodies (cdna4 + shared execute_shared.h), matching the codegen:
- v_cmp_class_f16 (vop3+vopc): classify from raw f16 bit fields instead of
  widening to f32 and using std::isnormal (an f16 denormal widens to an f32
  normal, so -denormal/+denormal classes 0x10/0x80 never matched).
- v_mov_b32 (vop3): apply the f32 abs/neg input modifiers (were dropped).
- pk u16 mul / int mad, i24 mul/mad, v_mul_lo_u16: widen to uint32/int64.
- v_bfe_i32: build masks with unsigned (1u<<w) and sign-extend via
  0xFFFFFFFFu<<w (was 1<<w / negation of INT_MIN -> UB).

rocjitsu_tests SIMD/vop suite: 209/209 pass; UBSan hits 23 -> 13 (remaining
are pre-existing shared/other-op shifts and muls, outside this change).

NOTE: a full `amdisa --multi --gen-shared-execute` regen is still pending
(needs the gfx1250 ISA spec); the generated bodies here were hand-applied to
match the codegen for verification.

+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/isa/arch/amdgpu/cdna4/vop3.cpp
@@ -2307,7 +2307,8 @@ void VMadI32I24Vop3::execute_impl(amdgpu::Wavefront &wf) {
     vdst.write_lane(wf, lane, [&]() {
       auto a = static_cast<int32_t>(src0.read_lane(wf, lane) << 8) >> 8;
       auto b = static_cast<int32_t>(src1.read_lane(wf, lane) << 8) >> 8;
-      return static_cast<uint32_t>(a * b + static_cast<int32_t>(src2.read_lane(wf, lane)));
+      return static_cast<uint32_t>(static_cast<int64_t>(a) * b +
+                                   static_cast<int32_t>(src2.read_lane(wf, lane)));
     }());
   }
 }
